Move interface bits to IIrrigation files
